+++ R/7_2_textTopicsAnalysis.R
@@ -1,13 +1,13 @@
 #' Wrapper for topicsTest function from the topics package
 #' @param model (list) The trained model
-#' @param pred_var_x (string) The x variable name to be predicted, and to be plotted (only needed for regression or correlation)
-#' @param pred_var_y (string) The y variable name to be predicted, and to be plotted (only needed for regression or correlation)
-#' @param group_var (string) The variable to group by (only needed for t-test)
-#' @param control_vars (vector) The control variables (not supported yet)
+#' @param x_variable (string) The x variable name to be predicted, and to be plotted (only needed for regression or correlation)
+#' @param y_variable (string) The y variable name to be predicted, and to be plotted (only needed for regression or correlation)
+# @param group_var (string) The variable to group by (only needed for t-test)
+#' @param controls (vector) The control variables (not supported yet)
 #' @param test_method (string) The test method to use, either "correlation","t-test", "linear_regression","logistic_regression", or "ridge_regression"
 # @param p_alpha (numeric) Threshold of p value set by the user for visualising significant topics
-# @param p_adjust_method (character) Method to adjust/correct p-values for multiple comparisons
-# (default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").
+#' @param p_adjust_method (character) Method to adjust/correct p-values for multiple comparisons
+#' (default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").
 # @param load_dir (string) The directory to load the test from, if NULL, the test will not be loaded
 # @param save_dir (string) The directory to save the test, if NULL, the test will not be saved
 #' @param ... Parameter settings from topicsTest in the topics-package.
@@ -16,20 +16,24 @@
 #' @export
 textTopicsTest <- function(
     model,
-    pred_var_x = NULL,
-    pred_var_y = NULL,
-    group_var = NULL, # only one in the case of t-test
-    control_vars = c(),
-    test_method = "linear_regression",
+    x_variable = NULL,
+    y_variable = NULL,
+    controls = c(),
+    test_method = "default",
+    p_adjust_method = "fdr",
     ...) {
 
   # Extract components from the model
   data <- model$train_data %>% tibble::as_tibble()
   preds <- model$preds %>% tibble::as_tibble()
   seed <- model$seed
-  save_dir <- model$save_dir
+#  save_dir <- model$save_dir
   ngrams <- model$ngrams
   model_summary <- model$model
+
+  # Adding dummy for prevalence and coherence to fit the topics package
+  model_summary$summary$prevalence <- NA
+  model_summary$summary$coherence <- NA
 
   # Call topicsTest from the topics package
   results <- topics::topicsTest(
@@ -37,18 +41,16 @@ textTopicsTest <- function(
     model = model_summary,
     preds = preds,
     ngrams = ngrams,
-    pred_var_x = pred_var_x,
-    pred_var_y = pred_var_y,
-    group_var = group_var,
-    control_vars = control_vars,
+    x_variable = x_variable,
+    y_variable = y_variable,
+#    group_var = group_var,
+    controls = controls,
     test_method = test_method,
-#    p_adjust_method = p_adjust_method,
-    seed = seed,
-    # load_dir = NULL,
-    save_dir = save_dir,
-    ...
+    p_adjust_method = p_adjust_method,
+    seed = seed
+    #, ...
   )
-
+  results
   return(results)
 }
 
@@ -142,44 +144,15 @@ textTopicsWordcloud <- function(
     test = NULL,
     save_dir,
     seed = 2024,
-#    color_scheme = "default",
-#    scale_size = FALSE,
-#    plot_topics_idx = NULL,
-#    save_dir,
-#    figure_format = "svg",
-#    width = 10,
-#    height = 8,
-#    max_size = 10,
-#    seed = 42,
-#    scatter_legend_dot_size = 15,
-#    scatter_legend_bg_dot_size = 9,
-#    scatter_legend_n = c(1,1,1,1,0,1,1,1,1),
-#    scatter_legend_method = c("mean"),
-#    scatter_legend_specified_topics = NULL,
-#    scatter_legend_topic_n = FALSE,
-#    grid_legend_title = "legend_title",
-#    grid_legend_title_size = 5,
-#    grid_legend_title_color = 'black',
-#    grid_legend_x_axes_label = "legend_x_axes_label",
-#    grid_legend_y_axes_label = "legend_y_axes_label",
-#    grid_legend_number_color = 'black',
-#    grid_legend_number_size = 5,
     ...) {
 
 
-  topics::topicsPlot(
+  plots <- topics::topicsPlot(
     model = model,
     ngrams = NULL,
     test = test,
-#    color_scheme = color_scheme,
-#    scale_size = scale_size,
-#    plot_topics_idx = plot_topics_idx,
     save_dir = save_dir,
-#    figure_format = figure_format,
-#    width = width,
-#    height = height,
-#    max_size = max_size,
-    seed = seed,
+    seed = seed
 #    scatter_legend_dot_size = scatter_legend_dot_size,
 #    scatter_legend_bg_dot_size = scatter_legend_bg_dot_size,
 #    scatter_legend_n = scatter_legend_n,
@@ -193,12 +166,12 @@ textTopicsWordcloud <- function(
 #    grid_legend_y_axes_label = grid_legend_y_axes_label,
 #    grid_legend_number_color = grid_legend_number_color,
 #    grid_legend_number_size = grid_legend_number_size,
-    ...
+    #, ...
   )
 
   completion_text <- paste0("The plots (p< p_alpha) are saved in ", save_dir, "/seed_", seed, "/wordclouds")
   message(colourise(completion_text, "green"))
 
+  return(plots)
 }
-